reimplement Java object monitors (second try)
See commit 8120bee4dc5f9ae2dec75a907778f1479ad398bd for the original
problem description and solution.  That commit and a couple of related
ones had to be reverted when we found they had introduced GC-safety
regressions leading to crashes.

This commit restores the reverted code and fixes the regressions.